requirements.txt
flask
pillow
numpy
pymilvus==1.1.2
requests
clip
